v0.29p2

+++ Python2D/util/GameTimer.cpp
@@ -119,7 +119,7 @@ float XCGameTimer::getFramePerSecond()
 	return fpsNow;
 }
 
-double XCGameTimer::getAccumlateTime()
+long double XCGameTimer::getAccumlateTime()
 {
 	return accumulateTime;
 }
